feat: new-component-searchInp
