+++ src/components/Home/index.jsx
@@ -1,10 +1,11 @@
-import { Header, Slider } from 'components';
+import { Header, SearchInp, Slider } from 'components';
 import React from 'react';
 
 export default function Home() {
   return (
     <div>
       <Header />
+      <SearchInp />
       <Slider />
     </div>
   );
