+++ src/components/index.js
@@ -3,3 +3,4 @@ export { default as Signup } from './Signup';
 export { default as Slider } from './Slider';
 export { default as Signin } from './Signin';
 export { default as Header } from './Header';
+export { default as SearchInp } from './SearchInp';
